Moved PrimitiveTypeInfo to serde-api
In order to move PrimitiveTypeInfo we need to move the PrimitiveTypeEntry as well.
PrimitiveTypeEntry depends on PrimitiveObjectInspectorUtils which cannot be pulled
into serde-api. Hence the static final maps are moved to PrimitiveEntry and we provide
static access methods to these maps along with the register method to add the key
value pairs in the maps
