+++ ql/src/java/org/apache/hadoop/hive/ql/udf/generic/GenericUDFReflect2.java
@@ -39,7 +39,7 @@ import org.apache.hadoop.hive.serde2.objectinspector.ObjectInspector;
 import org.apache.hadoop.hive.serde2.objectinspector.PrimitiveObjectInspector;
 import org.apache.hadoop.hive.serde2.objectinspector.primitive.PrimitiveObjectInspectorFactory;
 import org.apache.hadoop.hive.serde2.objectinspector.primitive.PrimitiveObjectInspectorUtils;
-import org.apache.hadoop.hive.serde2.objectinspector.primitive.PrimitiveObjectInspectorUtils.PrimitiveTypeEntry;
+import org.apache.hadoop.hive.serde2.objectinspector.primitive.PrimitiveTypeEntry;
 import org.apache.hadoop.hive.serde2.objectinspector.primitive.StringObjectInspector;
 import org.apache.hadoop.io.BooleanWritable;
 import org.apache.hadoop.io.BytesWritable;
@@ -105,9 +105,9 @@ public class GenericUDFReflect2 extends AbstractGenericUDFReflect {
     return returnOI;
   }
 
-  private PrimitiveObjectInspectorUtils.PrimitiveTypeEntry getTypeFor(Class<?> retType)
+  private PrimitiveTypeEntry getTypeFor(Class<?> retType)
       throws UDFArgumentException {
-    PrimitiveObjectInspectorUtils.PrimitiveTypeEntry entry =
+    PrimitiveTypeEntry entry =
         PrimitiveObjectInspectorUtils.getTypeEntryFromPrimitiveJavaType(retType);
     if (entry == null) {
       entry = PrimitiveObjectInspectorUtils.getTypeEntryFromPrimitiveJavaClass(retType);
